Add all files from server VPS
This rar file (server.rar) contains all the folders and files that must be uploaded to the VPS server for the application to work properly. There is a database and all the scripts for operating the online functions of the application. In the application code, you must also change the IP address to correspond to the new VPS server.